% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shs_create_app_data.R
\name{shs_create_app_data}
\alias{shs_create_app_data}
\title{Process SHS survey Excel dataset, output to Rds files for use in Shiny app}
\usage{
shs_create_app_data(
  destination_directory,
  source_data_directory,
  columns_to_remove,
  column_names_workbook_path,
  variable_names_workbook_path
)
}
\arguments{
\item{destination_directory}{\code{string}.
The path of the directory to save the output data to.}

\item{source_data_directory}{\code{string}.
The path of the directory containing source data and metadata in Excel format.}

\item{columns_to_remove}{\code{string}.
Unnecessary columns to remove from the dataset.}

\item{column_names_workbook_path}{\code{string}.
The path of an Excel file containing column renaming details.}

\item{variable_names_workbook_path}{\code{string}.
The path of an Excel file containing variable renaming details.}
}
\value{
\code{null}.
}
\description{
\code{shs_create_app_data} extracts data and metadata, removes unnecessary columns,
renames columns and variables, processes data into the correct format, and removes local
authority data from Scotland-only tables.
}
\examples{
\dontrun{
shs_create_app_data(destination_directory, source_data_directory, column_names_workbook_path, variable_names_workbook_path, columns_to_remove)
}

}
