% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shs_create_reports.R
\name{shs_create_reports}
\alias{shs_create_reports}
\title{Create Rmd reports for each chapter}
\usage{
shs_create_reports()
}
\value{
\code{null}.
}
\description{
\code{shs_create_reports} creates an Rmd report file for each topic in the SHS annual report data.
}
\examples{
\dontrun{
shs_create_reports()
}

}
