% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shs_extract_chapter_names.R
\name{shs_extract_chapter_titles}
\alias{shs_extract_chapter_titles}
\title{Extract chapter titles from Excel}
\usage{
shs_extract_chapter_titles()
}
\value{
\code{null}.
}
\description{
\code{shs_extract_chapter_titles} extracts data from Excel workbooks in a specified location,
and saves all sheets as .Rds files in a newly created location.
}
\examples{
shs_extract_chapter_titles()

}
