% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shs_extract_column_and_variable_names.R
\name{shs_extract_column_and_variable_names}
\alias{shs_extract_column_and_variable_names}
\title{Extract survey column and variable_names from Excel}
\usage{
shs_extract_column_and_variable_names()
}
\value{
\code{null}.
}
\description{
\code{shs_extract_column_and_variable_names} extracts survey metadata from Excel workbooks in a specified location,
and saves each sheet into a specified destination as an individual \code{.Rds} file.
}
\examples{
\dontrun{
shs_extract_column_and_variable_names()
}

}
