% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shs_extract_survey_data.R
\name{shs_extract_survey_data}
\alias{shs_extract_survey_data}
\title{Extract data from Excel workbooks}
\usage{
shs_extract_survey_data()
}
\value{
\code{null}.
}
\description{
\code{shs_extract_survey_data} extracts data from Excel workbooks in a specified location,
and saves all sheets as RData files in a newly created location.
}
\examples{
shs_extract_survey_data()

}
