% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shs_get_column_and_variable_names.R
\name{shs_get_column_and_variable_names}
\alias{shs_get_column_and_variable_names}
\title{Extract column and variable names from raw data}
\usage{
shs_get_column_and_variable_names(
  destination_directory,
  source_data_directory,
  columns_to_remove,
  existing_column_names_path = NULL,
  existing_variable_names_path = NULL
)
}
\arguments{
\item{source_data_directory}{\code{string}.
The path of the directory containing source data and metadata.}

\item{columns_to_remove}{\code{string}.
Unnecessary columns to remove from the dataset. If incorrect values provided, output file variable_names.xlsx will be populated incorrectly.}

\item{existing_column_names_path}{\code{string}.
A path to an existing Excel file with the same structure. Any display_name values will be copied to the new output file.}

\item{existing_variable_names_path}{\code{string}.
A path to an existing Excel file with the same structure. Any display_name values will be copied to the new output file.}

\item{top_level_directory}{\code{string}.
The path of the directory the app will be created in.}
}
\value{
\code{null}.
}
\description{
\code{shs_get_column_and_variable_names} extracts column and variable names from an SHS survey dataset,
and writes the output to Excel.
The output file 'column_names.xlsx' will contain all column names in the source files, minus any specified in
the \code{column_names_to_exclude} argument.
The output file 'variable_names.xlsx' will contain all unique values present in the second column of all tables in the raw data,
once columns names in \code{column_names_to_exclude} are removed.
The resulting Excel sheets will both have two columns: source_name and display_name. The display name value is the value
that will be displayed in the final Shiny app, these values can be added manually or by providing paths of existing,
populated sheets to the arguments \code{existing_column_names_path} and \code{existing_variable_names_path}.
}
\examples{
\dontrun{
shs_get_column_and_variable_names(top_level_directory, source_data_directory, columns_to_remove, existing_column_names_path, existing_variable_names_path)
}

}
