% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shs_get_variable_names.R
\name{shs_get_variable_names}
\alias{shs_get_variable_names}
\title{Get variable names from an extracted dataset}
\usage{
shs_get_variable_names(extracted_data_path, column_and_variable_path)
}
\arguments{
\item{column_and_variable_path}{\code{string}. The path to save the variable names Excel workbook to.}

\item{extracted_dataset_path}{\code{string}. The path of the directory containing extracted survey data.}
}
\value{
\code{null}.
}
\description{
\code{shs_get_variable_names} gets variable names from extracted SHS survey data in a specified location,
and saves them to an xlsx file, which can then be updated with names to display the annual report Shiny app.
}
\examples{
\dontrun{
shs_get_variable_names(extracted_dataset_path)
}

}
