% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shs_process_combine_multiple_years.R
\name{shs_process_combine_multiple_years}
\alias{shs_process_combine_multiple_years}
\title{Combine data for multiple years into one table}
\usage{
shs_process_combine_multiple_years(extracted_dataset_path)
}
\arguments{
\item{extracted_dataset_path}{\code{string}. The path of the directory a dataset has previously been extracted to, by \code{shs_extract_data}.}
}
\value{
\code{null}.
}
\description{
\code{shs_process_combine_multiple_years} combines extracted SHS survey data for a figure or table which is split over multiple years into a single dataframe.
The new dataframe is saved to an .Rds file, and the original .Rds files are deleted.
}
\examples{
shs_process_combine_multiple_years(extracted_dataset_path)

}
