% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shs_process_extracted_data.R
\name{shs_process_extracted_data}
\alias{shs_process_extracted_data}
\title{Process extracted data}
\usage{
shs_process_extracted_data(data_path)
}
\value{
\code{null}.
}
\description{
\code{shs_process_extracted_data} process extracted raw SHS survey data and metadata, amnd updates.
TODO: update this when process is defined.
}
\examples{
shs_process_extracted_data()

}
