% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shs_replace_na.R
\name{shs_replace_na}
\alias{shs_replace_na}
\title{Replace NA values in processed dataframes}
\usage{
shs_replace_na(data_file_path, incomplete_years)
}
\arguments{
\item{data_file_path}{\code{string}. A path to an Rds file to replace values in.}

\item{incomplete_years}{\code{character vector}. The path to a file containing design factor values.}
}
\value{
\code{null}.
}
\description{
\code{shs_replace_na} replaces NA values in processed dataframes with hyphens, or asterisks for incomplete years' data
}
\examples{
\dontrun{
shs_replace_na(dataframe, c(incomplete_year_1, incomplete_year_2))
}

}
