% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shs_shiny_variables.R
\name{shs_shiny_variables}
\alias{shs_shiny_variables}
\title{Create R variables to use in the Shiny App}
\usage{
shs_shiny_variables()
}
\value{
\code{null}.
}
\description{
\code{shs_shiny_variables} creates a file of R variables to based on extracted data, to be used in the SHS annual report Shiny app.
}
\examples{
\dontrun{
shs_shiny_variables()
}

}
