% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shs_update_column_names.R
\name{shs_update_column_names}
\alias{shs_update_column_names}
\title{Update column names from an extracted dataset}
\usage{
shs_update_column_names(new_workbook_path, old_workbook_path)
}
\arguments{
\item{new_workbook_path}{\code{string}. The path of the file containing the new column names.}

\item{old_workbook_path}{\code{string}. The path of the file containing the previous diplay name of the column.}
}
\value{
\code{null}.
}
\description{
\code{shs_update_column_names} updates column names in an  Excel spreadsheet with populated values from another spreadsheet.
}
\examples{
shs_update_column_names(new_workbook_path, old_workbook_path)

}
