% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shs_update_variable_names.R
\name{shs_update_variable_names}
\alias{shs_update_variable_names}
\title{Update variable names from an extracted dataset}
\usage{
shs_update_variable_names(new_workbook_path, old_workbook_path)
}
\arguments{
\item{new_workbook_path}{\code{string}. The path of the file containing the new variable names.}

\item{old_workbook_path}{\code{string}. The path of the file containing the previous diplay name of the variable.}
}
\value{
\code{null}.
}
\description{
\code{shs_update_variable_names} updates variable names in an  Excel spreadsheet with populated values from another spreadsheet.
}
\examples{
shs_update_variable_names(new_workbook_path, old_workbook_path)

}
