% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shs_create_app.R
\name{shs_create_app}
\alias{shs_create_app}
\title{Second part of SHS Data Explorer app processing}
\usage{
shs_create_app(
  destination_directory,
  data_directory,
  reports_start_year,
  reports_end_year
)
}
\arguments{
\item{destination_directory}{\code{string}.
The path of the directory to create the app in.}

\item{data_directory}{\code{string}.
The path of the directory containing source data and metadata.}

\item{reports_start_year}{\code{string}.
The first year available to select in the app's report generator.}

\item{reports_end_year}{\code{string}.
The final year available to select in the app's report generator.}
}
\value{
\code{null}.
}
\description{
\code{shs_create_app} creates an app that can be run/deployed, based on data created by
\code{shs_create_app_data}.
}
\examples{
\dontrun{
shs_create_app(destination_directory, data_directory, reports_start_year, reports_end_year)
}

}
