% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shs_extract_column_and_variable_names.R
\name{shs_extract_column_and_variable_names}
\alias{shs_extract_column_and_variable_names}
\title{Extract survey column and variable_names from Excel}
\usage{

  shs_extract_column_and_variable_names(source_column_and_variable_names_path,
  extracted_metadata_path)
}
\arguments{
\item{source_column_and_variable_names_path}{\code{string}. The path of the directory containing survey column and variable_names in Excel format.}

\item{extracted_metadata_path}{\code{string}. The path of the directory to be extracted to, created by \code{shs_extract_data}.}
}
\value{
\code{null}.
}
\description{
\code{shs_extract_column_and_variable_names} extracts survey metadata from Excel workbooks in a specified location,
and saves each sheet into a specified destination as an individual \code{.Rds} file.
}
\examples{
\dontrun{
shs_extract_column_and_variable_names(source_column_and_variable_names_path, extracted_metadata_path)
}

}
