% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shs_extract_data.R
\name{shs_extract_data}
\alias{shs_extract_data}
\title{Extract SHS data from Excel workbooks}
\usage{
shs_extract_data()
}
\value{
\code{null}.
}
\description{
\code{shs_extract_data} extracts raw SHS survey data and metadata from Excel workbooks in a specified location,
and saves all sheets as Rds files in a newly created location. Uses internal functions \code{shs_extract_dataset}
(to extract the source dataset) and \code{shs_extract_metadata} (to extract the source metadata).
}
\examples{
shs_extract_data()

}
