% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shs_get_column_names.R
\name{shs_get_column_names}
\alias{shs_get_column_names}
\title{Get column names from an extracted dataset}
\usage{
shs_get_column_names(extracted_data_path, column_and_variable_path)
}
\arguments{
\item{extracted_data_path}{\code{string}. The path of the directory containing extracted survey data.}

\item{column_and_variable_path}{\code{string}. The path to save the column names Excel workbook to.}
}
\value{
\code{null}.
}
\description{
\code{shs_get_column_names} gets column names from extracted SHS survey data in a specified location,
and saves them to an xlsx file, which can then be updated with names to display in the annual report Shiny app.
}
\examples{
\dontrun{
shs_get_column_names(extracted_dataset_path)
}

}
