% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shs_get_variable_names.R
\name{shs_get_variable_names}
\alias{shs_get_variable_names}
\title{Get variable names from an extracted dataset}
\usage{
shs_get_variable_names()
}
\value{
\code{null}.
}
\description{
\code{shs_get_variable_names} gets variable names from extracted SHS survey data in a specified location,
and saves them to an xlsx file, which can then be updated with names to display the annual report Shiny app.
}
\examples{
\dontrun{
shs_get_variable_names()
}

}
