% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shs_process_data.R
\name{shs_process_data}
\alias{shs_process_data}
\title{Process SHS annual report data for publication}
\usage{
shs_process_data(extracted_data_path)
}
\arguments{
\item{extracted_data_path}{\code{string}. The path of the directory data has previously been extracted to, by \code{shs_extract_data}.}
}
\value{
\code{null}.
}
\description{
\code{shs_process_data} processes raw SHS survey data which has been extracted from Excel sheets into individual .Rds files.
It relies on a number of internal functions: \code{shs_process_column_names}, TODO: update as more functions added.
}
\examples{
shs_process_data("extracted_data_path")

}
