% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shs_process_variable_names.R
\name{shs_process_variable_names}
\alias{shs_process_variable_names}
\title{Rename variable names in extracted datasets}
\usage{
shs_process_variable_names()
}
\value{
\code{null}.
}
\description{
\code{shs_process_variable_names} renames variables in extracted datasets according to data specified in \code{variable_names.Rds} in extracted metadata.
This metadata is extracted from an Excel sheet \code{variable_names.xlsx}. For more information see \code{shs_extract_data} and the internal function
\code{shs_extract_metadata}.
}
\examples{
\dontrun{
shs_process_variable_names()
}

}
