% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shs_remove_columns.R
\name{shs_remove_columns}
\alias{shs_remove_columns}
\title{Remove columns from extracted datasets}
\usage{
shs_remove_columns(extracted_data_path, columns_to_remove)
}
\arguments{
\item{columns_to_remove}{\code{character vector}. The names of all columns to remove.}

\item{extracted_dataset_path}{\code{string}. The path of the directory containing extracted survey data.}
}
\value{
\code{null}.
}
\description{
\code{shs_remove_columns} removes columns from extracted datasets.
}
\examples{
\dontrun{
shs_remove_columns(extracted_dataset_path, columns_to_remove)
}

}
