% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shs_rename_columns.R
\name{shs_rename_columns}
\alias{shs_rename_columns}
\title{Rename column names in Rda files}
\usage{
shs_rename_columns(extracted_dataset_path, extracted_metadata_path)
}
\value{
\code{null}.
}
\description{
\code{shs_rename_columns} renames column names in Rda files in extracted datasets according to data specified in extracted metadata.
}
\examples{
shs_rename_columns()

}
